address TODO about compilerOptions
